+++ src/constants.js
@@ -1,8 +1,4 @@
-const EVENTS = {
+export const EVENTS = {
   ATTACH_SERVER: 'ATTACH_SERVER',
   DETACH_SERVER: 'DETACH_SERVER',
-};
-
-module.exports = {
-  EVENTS,
 };